requirements.txt
pytest>=7.0
pyyaml>=6.0
pandas>=1.5
numpy>=1.23
scipy>=1.9
